% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pseudotime.R
\name{calc_pseudotime}
\alias{calc_pseudotime}
\title{Calculate pseudotime}
\usage{
calc_pseudotime(embedding, clusters)
}
\arguments{
\item{embedding}{data.frame}

\item{clusters}{factor}
}
\value{
data.frame
}
\description{
Calculate pseudotemporal order via principal curves for an arbitrary low dimensional embedding for scRNAseq data.
}
