% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_small.R
\docType{data}
\name{cycle.small}
\alias{cycle.small}
\title{Cycle embeddings}
\format{A data frame with 774 rows and 4 columns:
\describe{
  \item{PC_1}{principal component 1}
  \item{PC_2}{principal component 2}
  \item{PC_3}{principal component 3}
  \item{clusters}{cluster assignments}
}}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/27365425/}
}
\usage{
cycle.small
}
\description{
This dataset is a 2 dimensional PC embedding of cells progressing through the cell cycle, with some clusters assigned.
}
\keyword{datasets}
