% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_small.R
\docType{data}
\name{genes.small}
\alias{genes.small}
\title{Small expression matrix}
\format{A data frame with 65 rows and 744 columns:
\describe{
  \item{Prog_013}{cell expression profile}
  \item{Prog_019}{cell expression profile}
  ...
}}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/27365425/}
}
\usage{
genes.small
}
\description{
This dataset is subset of a gene expression matrix that we will use to demonstrate functionality of finding genes associated with pseudotime.
}
\keyword{datasets}
