% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_over_time.R
\name{genes_over_time}
\alias{genes_over_time}
\title{Fit genes as a function of pseudotime}
\usage{
genes_over_time(expr.matrix, pseudotime)
}
\arguments{
\item{expr.matrix}{data.frame}

\item{pseudotime}{numeric}
}
\value{
list
}
\description{
Utilize generalized additive models to fit gene expression as a function of pseudotime.
}
